feat: move operational solution loader to the correct file
